requirements.txt
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
config==0.5.0.post0
Cython==0.29.14
Flask==1.1.2
gensim==3.8.3
idna==2.10
itsdangerous==1.1.0
Janome==0.4.1
Jinja2==2.11.2
MarkupSafe==1.1.1
mecab-python3==1.0.3
numpy==1.19.4
psutil==5.7.3
requests==2.25.0
scipy==1.5.4
six==1.15.0
smart-open==3.0.0
tqdm==4.53.0
urllib3==1.26.2
Werkzeug==1.0.1
